Bug fix for date calculation
I had the minimum & maximum age declared as constants. But:

'A constant in Ruby is like a variable, except that its value is
supposed to remain constant for the duration of a program.'

They would never be updated after the initial load of the program. The
reason we haven't stumbled upon this bug just yet is because we deploy
pretty frequently.

Bug caught by Dan Palmer. All hail Dan.