Implement QuestionFormFactory
The factory takes care of instantiating a new form based on the given
id. This will nicely plug in the controller.

I think one step further will be required which is to take the list of
the questions off this class. Maybe into a separate one, because that
list should only exist in one place and the factory is not the right
one.