Increase reusability in form templates
The form partials were originally representing the template for the
whole page. Including the errors, header, form definition, etc. But
because a lot of it is the same for all the forms, it's now been taken
out to the main edit template and the form partials only contain the
form body itself.

There's another set of partials for each of the forms, which is a header
partial. Those contain the pre-form page content ie. examples. About
half of them are empty, but because there isn't a nice way in rails to
only render a partial if exists, I put those files as empty. Based on
user testing, we may add some text there anyway. If this proves to be
the final state, we may refactor this later to only store the non-empty
files somehow. But trying to render a partial and `rescue nil` doesn't
seem like a nice solution.