Create SessionsController to manipulate session
For the moment it can delete all data from the session. But I think we
could use the same controller on 'Apply now' as well, if we need to do
more about sessing it up. Then it could just redirect to the first
question, so that we don't create records on GET.