Redirect home from questions when storage is empty
If a user navigates to a question (apart from the very first one) and
their session is empty, they're redirected to the home page. We're not
expecting users to be changing url manually, but the situation could
occur when they use the browser back button after submission, when the
session is cleared.