Display list of selected kinds of income
It's a bit hacky solution to read from the session directly in the
template. But at the moment there's no way of using previous answers
other than the ones that have been exported to form OnlineApplication
model. And considering this is an anomaly that we ask a question where
we don't actually care about the answer, I think it's acceptable.